\name{clusterpathRcpp-package}
\alias{clusterpathRcpp-package}
\docType{package}
\title{Fast agglomerative convex clustering, original Rcpp implementation}
\description{Implements the clusterpath, a clustering method in which all
the clusters are solutions to a convex optimization problem with fusion
penalties. This package uses the original Rcpp bindings
to wrap the C++ implementation.}
\details{
\tabular{ll}{Package: \tab clusterpathRcpp\cr
Type: \tab Package\cr
Title: \tab Fast agglomerative convex clustering, original Rcpp implementation\cr
Version: \tab 1.1\cr
Date: \tab 2013-01-24\cr
Author: \tab Toby Dylan Hocking\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 2.11.0), Rcpp (>= 0.9.9), lattice, ggplot2, plyr, reshape2,
latticeExtra, Matrix, grid\cr
Suggests: \tab rgl, directlabels, mclust, multicore\cr
LinkingTo: \tab Rcpp\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


