\name{halfmoon.cluster}
\alias{halfmoon.cluster}
\title{halfmoon cluster}
\description{simulate some data in the shape of 2 half-moons and run some
clustering algorithms on them.}
\usage{halfmoon.cluster(seed, N = 150, noise = 0.5, ...)}
\arguments{
  \item{seed}{
}
  \item{N}{
}
  \item{noise}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





