\name{plot.predictclusterpath}
\alias{plot.predictclusterpath}
\title{plot predictclusterpath}
\description{Plot the result of a predictclusterpath data frame as points. }
\usage{\method{plot}{predictclusterpath}(x, type = "p", ...)}
\arguments{
  \item{x}{
}
  \item{type}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





